requirements.txt
flask
flask-cors
openai-whisper
torch
